requirements.txt
playwright>=1.40.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.2
